11-28
Todo: Singleton DB 연결, 검사 결과 class 추가

+++ factoryMethod/AnemiaScreening.java
@@ -4,11 +4,11 @@ import java.util.HashMap;
 
 public class AnemiaScreening implements Screening{ //빈혈 검사
 
-    HashMap<String, Integer> data = new HashMap<>();
+    HashMap<String, Double> data = new HashMap<>();
 
     @Override
-    public HashMap<String, Integer> getData() {
-        data.put("혈색소",0);
+    public HashMap<String, Double> getData() {
+        data.put("혈색소",0.0);
         return data;
     }
 }
