+++ factoryMethod/HighBloodPressureScreening.java
@@ -4,12 +4,12 @@ import java.util.HashMap;
 
 class HighBloodPressureScreening implements Screening { // 혈압 검사
 
-    HashMap<String, Integer> data = new HashMap<>();
+    HashMap<String, Double> data = new HashMap<>();
 
     @Override
-    public HashMap<String, Integer> getData() {
-        data.put("이완기",0);
-        data.put("수축기",0);
+    public HashMap<String, Double> getData() {
+        data.put("이완기",0.0);
+        data.put("수축기",0.0);
         return data;
     }
 }
