+++ factoryMethod/Screening.java
@@ -1,9 +1,10 @@
 package factoryMethod;
 
+import java.util.ArrayList;
 import java.util.HashMap;
 
 public interface Screening {
-    public HashMap<String, Integer> getData();
+    public HashMap<String, Double> getData();
 }
